LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/flow-content-0/dialog-expected.txt




FAIL Closed dialog in width: 540px iframe assert_equals: display expected "none" but got "inline"
FAIL Open dialog in width: 540px iframe assert_equals: display expected "block" but got "inline"
FAIL Modal dialog in width: 540px iframe assert_equals: display expected "block" but got "inline"
FAIL Closed dialog in width: 538px iframe assert_equals: display expected "none" but got "inline"
FAIL Open dialog in width: 538px iframe assert_equals: display expected "block" but got "inline"
FAIL Modal dialog in width: 538px iframe assert_equals: display expected "block" but got "inline"
